agents/requirements.txt
livekit-agents>=1.3.3
livekit-plugins-google>=1.3.3
livekit-plugins-deepgram>=1.3.3
livekit-plugins-cartesia>=1.3.3
livekit-plugins-silero>=1.3.3
livekit-plugins-anam>=1.3.3
python-dotenv>=1.0.1
httpx>=0.27.0
